Pycharm refactor failed me.
