+++ allianceauth/services/modules/discourse/tasks.py
@@ -3,7 +3,7 @@ import logging
 from django.contrib.auth.models import User
 from django.core.exceptions import ObjectDoesNotExist
 
-from allianceauth.celeryapp import app
+from allianceauth.celery import app
 from allianceauth.notifications import notify
 from .manager import DiscourseManager
 from .models import DiscourseUser
